Fix: warning fixes
Former-commit-id: c2628cb07ee9e1c15394c2b733a1d84748f2138a

+++ Sources/Internal/Entity/EntityManager.cpp
@@ -36,7 +36,7 @@ void EntityManager::AddComponent(Entity * entity, Component * component)
         uint64 bit = newFamilyType.GetBit();
         for (uint64 idx = 0; idx < 64; ++idx)
         {
-            if (bit & (1 << idx))
+            if (bit & ((int64)1 << idx))
             {
                 Component * comp = Component::GetComponentByIndex(idx);
                 familiesWithComponent.insert(std::pair<Component*, EntityFamily*>(comp, newFamily));
